requirements.txt
hyperliquid-python-sdk
requests
python-dotenv
